+++ src/migrations/Install.php
@@ -10,13 +10,19 @@
 
 namespace mortscode\feedback\migrations;
 
+use craft\elements\Entry;
+use craft\errors\ElementNotFoundException;
+use mortscode\feedback\enums\FeedbackOrigin;
 use mortscode\feedback\enums\FeedbackStatus;
 use mortscode\feedback\enums\FeedbackType;
-use mortscode\feedback\Feedback;
 
 use Craft;
-use craft\config\DbConfig;
 use craft\db\Migration;
+use mortscode\feedback\Feedback;
+use mortscode\feedback\helpers\RatingsHelpers;
+use Throwable;
+use yii\base\Exception;
+use yii\base\ExitException;
 
 /**
  * Feedback Install Migration
@@ -65,6 +71,7 @@ class Install extends Migration
             // Refresh the db schema caches
             Craft::$app->db->schema->refresh();
             $this->insertDefaultData();
+            $this->updateEntryRatings();
         }
 
         return true;
@@ -84,6 +91,7 @@ class Install extends Migration
     {
         $this->driver = Craft::$app->getConfig()->getDb()->driver;
         $this->removeTables();
+        $this->resetEntryRatings();
 
         return true;
     }
@@ -130,7 +138,11 @@ class Install extends Migration
                         FeedbackStatus::Spam,
                     ]),
                     'PRIMARY KEY(id)',
-                    'isImport' => $this->boolean(),
+                    'feedbackOrigin' => $this->enum('feedbackOrigin', [
+                        FeedbackOrigin::FRONTEND,
+                        FeedbackOrigin::CONTROL_PANEL,
+                        FeedbackOrigin::IMPORT_DISQUS,
+                    ]),
                 ]
             );
         }
@@ -178,7 +190,57 @@ class Install extends Migration
      */
     protected function removeTables(): void
     {
-    // feedback_record table
+        // drop feedback_record tables
         $this->dropTableIfExists('{{%feedback_record}}');
+    }
+
+    /**
+     * @throws Throwable
+     * @throws ElementNotFoundException
+     * @throws Exception
+     * @throws ExitException
+     */
+    protected function resetEntryRatings(): void
+    {
+        // remove all ratings from entries
+        $entries = Entry::findAll();
+
+        foreach ($entries as $entry) {
+            if ($entry) {
+                if(isset($entry->averageRating)){
+                    $entry->setFieldValue('averageRating', 'null');
+                }
+                if(isset($entry->totalRatings)){
+                    $entry->setFieldValue('totalRatings', 0);
+                }
+                if(isset($entry->totalPending)){
+                    $entry->setFieldValue('totalPending', 0);
+                }
+
+                Craft::$app->elements->saveElement($entry, false, true, false);
+            }
+        }
+    }
+
+    /**
+     * @throws Throwable
+     * @throws ElementNotFoundException
+     * @throws Exception
+     * @throws ExitException
+     */
+    protected function updateEntryRatings(): void
+    {
+        // remove all ratings from entries
+        $entries = Entry::findAll();
+
+        foreach ($entries as $entry) {
+            if ($entry) {
+                $entry->setFieldValue('averageRating', RatingsHelpers::getAverageRating($entry->id));
+                $entry->setFieldValue('totalRatings', RatingsHelpers::getTotalRatings($entry->id));
+                $entry->setFieldValue('totalPending', RatingsHelpers::getTotalPending($entry->id));
+
+                Craft::$app->elements->saveElement($entry, false, true, false);
+            }
+        }
     }
 }