+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.2 - 2021-06-08
+### Added
+- Element Property for FeedbackOrigin (Frontend, CP, or Import from disqus)
+- Ratings values update much better
+
 ## 1.0.1 - 2021-06-04
 ### Added
 - Email integration